Move storage data types to separate module
